Setup i18next
Currently only used in SummaryStats.
Also changed to `json` instead of `key_value_json` for the POEditor
export.
The whole `i18n` directory is now deleted when updating translations,
so that languages which have fallen behind are not kept in the folder.

Plurals are hard. For languages with only two plurals, the singular
is the plain key and the plural has a suffix of `_plural`. For
languages with multiple plurals, such as Arabic, all keys have a
suffix from `_0` to `_5` depending on the type. The update-translations
script has to take this into account. I experimented with the
`simplifyPluralSuffix` option, but it causes languages with only two
forms to have `_0` be the singular and `_1` be the plural, which is not
consistent with the suffixes of other languages (Arabic has a zero form,
which is `_0` and the singular is `_1`).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>