Use server-side cursor pagination in the Query Log
Also, the Query Log has been updated to use the new log format which
comes with the addition of pagination.

A new column, Reply, has been added. It holds the reply type and
response time.

Filtering and sorting are disabled until they can be reimplemented.

The translations have been updated, as a few strings were added.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>