WithAPIData: Don't make a request on refresh if we know the answer
If a user changes a setting, we send it to the API, and it says the
change was successful, then we should not have to ask the API for the
settings we just saved.

This is the case for status and preferences. When we save those
settings, we can update our local data with the new setting instead of
making a request for it. This changes a "PUT then GET" flow into a
"PUT then update local" flow, which is one less server call.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>