Use the generated list of languages to avoid 404 errors
If the user's language is detected to be a language not supported (for
example, en-US), then previously it would try to load the translation
files for that language, and fail with 404s. Now it will only use the
language if it is supported (instead of en-US, it will use en because
that is in the list).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>