Support refreshing on GitHub Pages
GitHub Pages does not support client-side routing, so when a user is on
`/whitelist` and refreshes, they get a 404 because there is no
`whitelist` file on the server. This is fixed by adding a custom 404
page which handles redirecting to index.html with the original path.

Credit for this goes to https://github.com/rafrex/spa-github-pages

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>