Add time range selector to dashboard and integrate summary stats
The time range selector shows the currently selected time range (uses
the label for predefined ranges).

The summary stats will change to using the database call when the
time range is not "Last 24 Hours". Also, when using the database call it
will only make the call once instead of updating periodically.

The date ranges were taken from the Query Log, so it has also been
updated to default to the last 24 hours. This makes more sense than
using "Today", as it now matches the dashboard's default range.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>