Load new pages in the Query Log faster
Previously each time data needed to be fetched for the query log, it was
debounced for 350ms. This includes fetching new pages when the user
clicks "Next", which causes unnecessary slowdowns. The actual time it
takes to fetch a new page is usually around 20ms, but when it's
debounced it takes 370ms.

The logic is that if the filters have not changed, then don't debounce
the call. The debounce mechanism is only there to prevent multiple calls
when the filters are in flux (user is typing in the search box).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>