Fix sending the authentication request twice on the login screen
If you click "Enter" on your keyboard, it will fire off a form submitted
event which will run the authentication code. However, it will run the
authentication again with a possibly password due to the key down event
handler. This event handler is no longer needed because we have the form
submit handler.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>