Fix debouncing in query log before selecting filters
We populated the state filters with a filter (last 24 hours), but did
not do the same for the table filters, as they are handled internally by
the table. This caused all requests to debounce, such as loading a new
page, until a filter was modified.

Now the filters are aligned from the beginning. Also, we use a better
filter comparison method (lodash's `isEqual`), as using `===` does not
do a good job with arrays in JavaScript (ex. `[] !== []`).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>