Improve CI test performance by limiting the number of workers
Around the time of merging #264 and afterwards, the CI would sometimes
freeze while running the tests. When SSHing into the box with the tests
running, several node processes were running. Jest also recommends to
limit the number of workers if CI performance is lacking:

https://jestjs.io/docs/en/troubleshooting.html#tests-are-extremely-slow-on-docker-and-or-continuous-integration-ci-server

This change is applied in this commit.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>