Auto-login the user if no password is required
See pi-hole/api#173 for more details.

This implementation refreshes the page if it was started in a logged out
state and it is found that the API doesn't require authentication. It is
explained in more detail in a code comment, but we should look into
Redux to make this and many other things cleaner and more testable.

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>