Add Redux Saga
Redux Saga provides a clean and testable way to run side effects while
integrating with Redux. It is used now to fetch the preferences data and
store it in the Redux store, and will later be used for more side
effects (like updating the preferences in the API).

The Redux store React provider component is moved to the top level. The
other middleware used in development mode comes from the default
middleware. We are not using `getDefaultMiddleware` because it includes
Redux Thunk, which we do not need (we use Saga instead).

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>