Fix auto-login always refreshing in fake API mode by migrating to a Saga
The original auto-login code was written before Redux and was planned to
be migrated. This migrates the code into a Redux Saga and adds a test
suite. In the process, #365 is fixed.

In the future, the logged in state will be distributed via Redux, but
until then we still have to refresh the page. The saga continues the
refreshing behavior, but it will be easy to change this to emit Redux
store updates once the logged in state moves to Redux.

One test is implemented sub-optimally due to a bug in the saga test
library:
https://github.com/jfairbank/redux-saga-test-plan/issues/296
I have proposed a fix here:
https://github.com/jfairbank/redux-saga-test-plan/pull/298

Fixes #365

Signed-off-by: Mcat12 <newtoncat12@yahoo.com>